requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.5
cffi==1.14.0
Django==3.0.7
django-appconf==1.0.4
django-imagekit==4.0.2
djangorestframework==3.11.0
dnspython==1.16.0
future==0.18.2
gevent==20.6.0
greenlet==0.4.16
kombu==4.6.10
monotonic==1.5
pilkit==2.0
Pillow==7.1.2
pycparser==2.20
pytz==2020.1
redis==3.5.3
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
vine==1.3.0
zope.event==4.4
zope.interface==5.1.0
